[feat] 회원, 마이페이지 관련 기능 구현 및 팝업 추가
- 회원가입, 마이페이지 및 마이리스트 관련 코드 구현
- 마이리스트의 디테일 페이지에서 상세 팝업 구현
- 컨펌 창 및 얼럿 창 기능 추가하여 사용자 확인 처리
- 마이페이지에서 일정 상세보기 팝업 활성화
- 상태 관리 및 사용자 인터페이스 개선

+++ src/components/mypage/MyplanlistItem.jsx
@@ -1,27 +1,26 @@
-import { Link, useNavigate } from "react-router-dom";
-import {} from "./mypage";
-
-import { IoClose } from "react-icons/io5";
-import { DeleteBtn } from "../../pages/mypage/plan-list/myplan";
-import axios from "axios";
 import { useContext } from "react";
+import { Link } from "react-router-dom";
 import { LoginContext } from "../../contexts/LoginContext";
+import axios from "axios";
+// styled
+import { DeleteBtn } from "../../pages/mypage/plan-list/myplan";
+// icon
+import { IoClose } from "react-icons/io5";
 
-const MyplanlistItem = ({ item, myScheduleList, setMyScheduleList }) => {
+const MyplanlistItem = ({ item, setMyScheduleList }) => {
   const planDateAddOne = parseInt(item?.planDate) + 1;
   const imgUrl = `http://112.222.157.156:5212/pic/city/${item?.cityId}/${item?.cityPic}`;
   const { user } = useContext(LoginContext);
 
   const deleteItem = _Id => {
-    // const arr = myScheduleList.filter(item => item.planMasterId !== _Id);
     setMyScheduleList(prev => prev.filter(item => item.planMasterId !== _Id));
-    // navigate("/auth");
   };
   // http://112.222.157.156:5212/api/plan?planMasterId=0&userId=0
   const deleteListItem = async data => {
     try {
       const res = await axios.delete(
         `/api/plan?planMasterId=${data}&userId=${user.userId}`,
+        alert("일정이 삭제되었습니다."),
       );
       console.log(res);
     } catch (error) {
@@ -33,8 +32,13 @@ const MyplanlistItem = ({ item, myScheduleList, setMyScheduleList }) => {
 
   const handleClickbtn = e => {
     e.preventDefault();
-    deleteItem(item.planMasterId);
-    deleteListItem(item.planMasterId);
+    const isConfirm = confirm("선택한 일정을 삭제하시겠습니까?");
+    if (isConfirm) {
+      deleteItem(item.planMasterId);
+      deleteListItem(item.planMasterId);
+    } else {
+      console.log("일정 삭제 취소됨");
+    }
   };
 
   return (
